Refactor and remove necessary resources from .gitignore
