+++ src/Resources/images.qrc
@@ -1,0 +1,10 @@
+<RCC>
+    <qresource prefix="/">
+        <file>Resources/statDEX.png</file>
+        <file>Resources/statFTH.png</file>
+        <file>Resources/statMAG.png</file>
+        <file>Resources/statSTR.png</file>
+        <file>Resources/statTEC.png</file>
+        <file>Resources/statVIT.png</file>
+    </qresource>
+</RCC>
