+++ src/headers/strings.h
