+++ src/sources/mainwindow.cpp
@@ -1,4 +1,4 @@
-#include "mainwindow.h"
+#include "../headers/mainwindow.h"
 
 MainWindow::MainWindow(QWidget *parent) :
     QMainWindow(parent),
